requirements.txt
streamlit
langchain
faiss-cpu
PyMuPDF
langchain_community
langchain-huggingface
langchain_ollama